refactor(tailer): move rotation handling into rotation_detect.go
CodeScene flagged the #1104 fix for Low Cohesion (critical) on
tailer.go and Complex Method on TailAndProcess. Both are fair: the
change piled a fingerprint window, three helpers and a switch arm into
an already-large file, and grew the one function every pass runs
through.

Split along the seam the package already uses (tailer_ledger.go,
tailer_metrics.go): rotation_detect.go now owns deciding whether the
resume point is still valid and clearing what belongs to the previous
contents. resetAccumulatorsForRotation moves with it -- it is the other
half of the same concern, and rotation_reset_test.go finally has a
source counterpart.

TailAndProcess drops from a 3-arm switch to `startPos :=
t.resolveStartPos(file, fileSize)`, which also carries #1088's long
transcript-prune rationale out of the hot function and onto the branch
it actually explains.

Pure code motion plus one extracted method -- no behavior change; the
#1104 regression tests were re-confirmed to still fail when the
detection arm is reverted.
